First test to pass.
