+++ .gitignore
@@ -1,3 +1,3 @@
 .idea
-/dist/
-tests/__pycache__/
+dist
+*__pycache__*
